django_test_tasks/old_django_test_tasks/requirements.txt
beautifulsoup4==4.6.3
certifi==2018.8.24
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.1
django-debug-toolbar==1.10.1
django-filter==2.0.0
django-money==0.14.3
djangorestframework==3.8.2
dry-rest-permissions==0.1.10
envparse==0.2.0
idna==2.7
lxml==4.2.5
Markdown==3.0.1
oauthlib==3.1.0
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
py-moneyed==0.7.0
pytz==2018.5
requests==2.19.1
six==1.11.0
sqlparse==0.2.4
Unidecode==1.0.22
urllib3==1.23
django-ckeditor==5.9.0
